0. Print list 04

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int idx = 0;
+	size_t idx = 0;
 
 	while (h != NULL)
 	{
